+++ dpaste/urls/__init__.py
@@ -1,11 +1,11 @@
-from django.conf.urls import include, url
+from django.urls import include, re_path
 
 urlpatterns = [
-    url(r"^", include("dpaste.urls.dpaste_api")),
-    url(r"^", include("dpaste.urls.dpaste")),
-    url(r"^i18n/", include("django.conf.urls.i18n")),
+    re_path(r"^", include("dpaste.urls.dpaste_api")),
+    re_path(r"^", include("dpaste.urls.dpaste")),
+    re_path(r"^i18n/", include("django.conf.urls.i18n")),
 ]
 
 # Custom error handlers which load `dpaste/<code>.html` instead of `<code>.html`
-handler404 = "dpaste.views.page_not_found"
-handler500 = "dpaste.views.server_error"
+handler404 = "dpaste.views.handler404"
+handler500 = "dpaste.views.handler500"