+++ dpaste/urls/dpaste.py
@@ -1,6 +1,6 @@
 from django.apps import apps
 from django.conf import settings
-from django.conf.urls import url
+from django.urls import re_path
 from django.views.decorators.cache import cache_control
 from django.views.decorators.clickjacking import xframe_options_exempt
 from django.views.generic import TemplateView
@@ -11,8 +11,8 @@ L = getattr(settings, "DPASTE_SLUG_LENGTH", 4)
 config = apps.get_app_config("dpaste")
 
 urlpatterns = [
-    url(r"^$", views.SnippetView.as_view(), name="snippet_new"),
-    url(
+    re_path(r"^$", views.SnippetView.as_view(), name="snippet_new"),
+    re_path(
         r"^about/$",
         cache_control(max_age=config.CACHE_TIMEOUT)(
             TemplateView.as_view(
@@ -22,18 +22,18 @@ urlpatterns = [
         ),
         name="dpaste_about",
     ),
-    url(r"^history/$", views.SnippetHistory.as_view(), name="snippet_history"),
-    url(
+    re_path(r"^history/$", views.SnippetHistory.as_view(), name="snippet_history"),
+    re_path(
         r"^(?P<snippet_id>[a-zA-Z0-9]{%d,})/?$" % L,
         views.SnippetDetailView.as_view(),
         name="snippet_details",
     ),
-    url(
+    re_path(
         r"^(?P<snippet_id>[a-zA-Z0-9]{%d,})/raw/?$" % L,
         views.SnippetRawView.as_view(),
         name="snippet_details_raw",
     ),
-    url(
+    re_path(
         r"^(?P<snippet_id>[a-zA-Z0-9]{%d,})/slim/?$" % L,
         xframe_options_exempt(
             views.SnippetDetailView.as_view(template_name="dpaste/details_slim.html")