+++ dpaste/urls/dpaste_api.py
@@ -1,8 +1,8 @@
-from django.conf.urls import url
+from django.urls import re_path
 from django.views.decorators.csrf import csrf_exempt
 
 from ..views import APIView
 
 urlpatterns = [
-    url(r"^api/$", csrf_exempt(APIView.as_view()), name="dpaste_api_create_snippet",)
+    re_path(r"^api/$", csrf_exempt(APIView.as_view()), name="dpaste_api_create_snippet",)
 ]